src/api/requirements.txt
authlib
azure-functions
azure-identity
azure-storage-blob
azure-storage-queue
azure-data-tables
azure-monitor-opentelemetry
cryptography==43.0.*